change U-Boot image to run uncached
U-Boot relies on the size of the image to achieve cache coherence, ie. the
image must be large enough so that after it has been written completely, its
entry point has been written back. this should probably be fixed in U-Boot.
as a workaround, load and run the bootloader in uncached memory.