fix a few bugs in xmodem code, and make timing more aggressive
it was expecting blocks to start at number 2. that obviously isn't any good.

it was always returning EOF, and therefore never acknowledged anything.

it didn't even try to ACK the EOF. it cannot guarantee to do so because it
may no longer be around when the retried EOT arrives, but it can at least
try to ack properly.

there was a 10 second timeout for each 128-byte block. as the image is meant
to be running at 38400 baud, this is ridiculous. 1s is just as good, and
makes transfers start faster.