hide stub in header of 0ECW image
this only works for nanoboot which puts the header before the actual image,
and for U-Boot because the uImage can declare this layout too. this means
that nanoboot can be set up to either load U-Boot or WinCE (but not both)
by placing an appropriate image on flash. it can also always load WinCE
over Xmodem, though downloading the 4MB WinCE0 takes almost 20 minutes
at 38.4kbaud.

FlashReader puts the header elsewhere and probably cannot boot this image.
or it might jump directly and successfully to WinCE; it's the "proper"
bootloader after all.