implement fallback to next flash image and dynamic load addresses
when an image returns or fails to load, keep reading from flash until a
working image is found, or the end of flash is reached. this allows an
extra level of failsafe by having several boot images in flash so that
boot still succeeds if the first boot image fails.

the "reserved" header word is now interpreted as the load address, and
the image is started by jumping to its load address. this is more flexible
than variable entry points only.

also refactor the code significantly, improving readability and size.