support for compressed images
they need to include their own decompressor, because nanoboot has no space
for several KB of decompression code. the decompressor and compressed image
are loaded to 0x80420000, ie. right after nanoboot itself, so that the
uncompressed image can be placed anywehere in the first 64MB of memory.
currently, for U-Boot, the load address is hardcoded to 0x80100000, but the
code supports arbitrary load addresses and entry points.